move pyspcbridge/lib to pyspcbridge as some tools apparently treats directories named "lib" special
