+++ pyspcbridge/exceptions.py
