+++ pyspcbridge/panel.py
@@ -1,7 +1,7 @@
 import logging
 
 from .const import ArmMode
-from .lib.spc_error import SpcError
+from .spc_error import SpcError
 
 _LOGGER = logging.getLogger(__name__)
