+++ pyspcbridge/spc_error.py
