+++ pyspcbridge/websocket.py
